Remove old matcher
